+++ gsa/src/web/pages/audits/__tests__/actions.js
@@ -40,13 +40,15 @@ describe('Audit Actions tests', () => {
   console.error = () => {};
 
   test('should render', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.new,
-      alterable: '0',
-      last_report: {report: {_id: 'id'}},
-      permissions: {permission: [{name: 'everything'}]},
-      target: {_id: 'id', name: 'target'},
-      usage_type: 'audit',
+      alterable: false,
+      reports: {
+        lastReport: {id: 'id'},
+      },
+      permissions: [{name: 'everything'}],
+      target: {id: 'id', name: 'target'},
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();
@@ -79,13 +81,15 @@ describe('Audit Actions tests', () => {
   });
 
   test('should call click handlers', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.done,
-      alterable: '0',
-      last_report: {report: {_id: 'id'}},
-      permissions: {permission: [{name: 'everything'}]},
-      target: {_id: 'id', name: 'target'},
-      usage_type: 'audit',
+      alterable: false,
+      reports: {
+        lastReport: {id: 'id'},
+      },
+      permissions: [{name: 'everything'}],
+      target: {id: 'id', name: 'target'},
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();
@@ -150,13 +154,15 @@ describe('Audit Actions tests', () => {
   });
 
   test('should not call click handlers without permissions', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.done,
-      alterable: '0',
-      last_report: {report: {_id: 'id'}},
-      permissions: {permission: [{name: 'get_task'}]},
-      target: {_id: 'id', name: 'target'},
-      usage_type: 'audit',
+      alterable: false,
+      reports: {
+        lastReport: {id: 'id'},
+      },
+      permissions: [{name: 'get_task'}],
+      target: {id: 'id', name: 'target'},
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();
@@ -233,13 +239,15 @@ describe('Audit Actions tests', () => {
   });
 
   test('should not call click handlers for stopped audit without permissions', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.stopped,
-      alterable: '0',
-      last_report: {report: {_id: 'id'}},
-      permissions: {permission: [{name: 'get_task'}]},
-      target: {_id: 'id', name: 'target'},
-      usage_type: 'audit',
+      alterable: false,
+      reports: {
+        lastReport: {id: 'id'},
+      },
+      permissions: [{name: 'get_task'}],
+      target: {id: 'id', name: 'target'},
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();
@@ -318,13 +326,15 @@ describe('Audit Actions tests', () => {
   });
 
   test('should not call click handlers for running audit without permissions', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.running,
-      alterable: '0',
-      last_report: {report: {_id: 'id'}},
-      permissions: {permission: [{name: 'get_task'}]},
-      target: {_id: 'id', name: 'target'},
-      usage_type: 'audit',
+      alterable: false,
+      reports: {
+        lastReport: {id: 'id'},
+      },
+      permissions: [{name: 'get_task'}],
+      target: {id: 'id', name: 'target'},
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();
@@ -400,13 +410,13 @@ describe('Audit Actions tests', () => {
   });
 
   test('should call click handlers for running audit', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.running,
-      alterable: '0',
-      in_use: true,
-      permissions: {permission: [{name: 'everything'}]},
-      target: {_id: 'id', name: 'target'},
-      usage_type: 'audit',
+      alterable: false,
+      inUse: true,
+      permissions: [{name: 'everything'}],
+      target: {id: 'id', name: 'target'},
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();
@@ -469,13 +479,15 @@ describe('Audit Actions tests', () => {
   });
 
   test('should call click handlers for stopped audit', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.stopped,
-      alterable: '0',
-      last_report: {report: {_id: 'id'}},
-      permissions: {permission: [{name: 'everything'}]},
-      target: {_id: 'id', name: 'target'},
-      usage_type: 'audit',
+      alterable: false,
+      reports: {
+        lastReport: {id: 'id'},
+      },
+      permissions: [{name: 'everything'}],
+      target: {id: 'id', name: 'target'},
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();
@@ -540,13 +552,15 @@ describe('Audit Actions tests', () => {
   });
 
   test('should disable report download if grc format is not defined', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.stopped,
-      alterable: '0',
-      last_report: {report: {_id: 'id'}},
-      permissions: {permission: [{name: 'everything'}]},
-      target: {_id: 'id', name: 'target'},
-      usage_type: 'audit',
+      alterable: false,
+      reports: {
+        lastReport: {id: 'id'},
+      },
+      permissions: [{name: 'everything'}],
+      target: {id: 'id', name: 'target'},
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();
@@ -587,18 +601,20 @@ describe('Audit Actions tests', () => {
   });
 
   test('should render schedule icon if task is scheduled', () => {
-    const audit = Audit.fromElement({
+    const audit = Audit.fromObject({
       status: AUDIT_STATUS.stopped,
-      alterable: '0',
-      last_report: {report: {_id: 'id'}},
-      permissions: {permission: [{name: 'everything'}]},
-      target: {_id: 'id', name: 'target'},
-      schedule: {
-        _id: 'schedule1',
-        name: 'schedule1',
-        permissions: {permission: [{name: 'everything'}]},
+      alterable: false,
+      reports: {
+        lastReport: {id: 'id'},
       },
-      usage_type: 'audit',
+      permissions: [{name: 'everything'}],
+      target: {id: 'id', name: 'target'},
+      schedule: {
+        id: 'schedule1',
+        name: 'schedule1',
+        permissions: [{name: 'everything'}],
+      },
+      usageType: 'audit',
     });
 
     const handleAuditClone = jest.fn();